[chore][pkg/stanza] Cleanup noop operator files
