+++ pkg/stanza/operator/transformer/noop/transformer_test.go
@@ -14,20 +14,6 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/testutil"
 )
 
-func TestBuildValid(t *testing.T) {
-	cfg := NewConfigWithID("test")
-	op, err := cfg.Build(testutil.Logger(t))
-	require.NoError(t, err)
-	require.IsType(t, &Transformer{}, op)
-}
-
-func TestBuildInvalid(t *testing.T) {
-	cfg := NewConfigWithID("test")
-	_, err := cfg.Build(nil)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "build context is missing a logger")
-}
-
 func TestProcess(t *testing.T) {
 	cfg := NewConfigWithID("test")
 	cfg.OutputIDs = []string{"fake"}
